requirements/base.txt
pytz==2021.1  # https://github.com/stub42/pytz
python-slugify==5.0.2  # https://github.com/un33k/python-slugify
Pillow==9.0.1  # https://github.com/python-pillow/Pillow
argon2-cffi==20.1.0  # https://github.com/hynek/argon2_cffi
redis==3.5.3  # https://github.com/andymccurdy/redis-py
hiredis==1.1.0  # https://github.com/redis/hiredis-py
celery==4.4.6  # pyup: < 5.0,!=4.4.7  # https://github.com/celery/celery
django-celery-beat==2.2.0  # https://github.com/celery/django-celery-beat
flower==0.9.7  # https://github.com/mher/flower
whitenoise==5.2.0  # https://github.com/evansd/whitenoise
rcssmin==1.0.6  # https://github.com/ndparker/rcssmin

# Django
# ------------------------------------------------------------------------------
django==3.1  # pyup: < 3.2  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.1.1  # https://github.com/jazzband/django-model-utils
django-crispy-forms==1.11.2  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.12.1  # https://github.com/jazzband/django-redis
# Django REST Framework
djangorestframework==3.12.4  # https://github.com/encode/django-rest-framework
django-cors-headers==3.7.0 # https://github.com/adamchainz/django-cors-headers
django-compressor==2.4.1  # https://github.com/django-compressor/django-compressor
django-admin-interface==0.16.3
phonenumbers==8.12.31
responses==0.14.0
WooCommerce==3.0.0
pandas==1.3.4
openpyxl==3.0.9
drf-yasg==1.20.0
model-bakery==1.3.3
django-anymail[mailgun]
djangorestframework-api-key==2.1.0
